feat: prefill chat with prompt
